# contracts/openapi/swagger.yaml
basePath: /api/v1
...
info:
  contact:
    email: felipe.lechitz@gmail.com
    name: Lechitz
    url: https://github.com/lechitz
  description: Public and internal REST API for aion-api. Swagger (OpenAPI 3.x) is
    generated via swaggo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lechitz/aion-api
  title: aion-api — REST API Documentation
  version: 0.0.1
paths:
...
  /chat/audio:
    post:
      consumes:
      - multipart/form-data
      description: Sends an audio file to be transcribed and processed by the AI assistant.
        Requires authentication.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Audio file (webm, wav, mp3, max 10MB, max 60s)
        in: formData
        name: audio
        required: true
        type: file
      - description: Language code (pt, en, es) or auto-detect if empty
        in: formData
        name: language
        type: string
      - description: 'Optional runtime provider override (for example: openai, ollama)'
        in: formData
        name: provider
        type: string
      - description: Optional runtime model override when provider is present
        in: formData
        name: model
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Voice chat response with transcription and AI response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid audio file or validation error
          schema:
            type: string
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            type: string
        "413":
          description: Audio file too large
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
        "503":
          description: Service unavailable - AI service is down
          schema:
            type: string
      security:
      - BearerAuth.: []
      summary: Send voice chat message
      tags:
      - ChatText
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {token}" (JWT).
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: Session-based authentication cookie (e.g., Redis-backed).
    in: cookie
    name: auth_token
    type: apiKey
    x-graphqlplayground: /graphql/ // Cross-reference to GraphQL Playground (not covered
      by Swagger)
swagger: "2.0"